third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/source-frame-expected.txt

Tests that it's possible to set breakpoint in source frame, and that source frame displays breakpoints and console errors.


Running: testConsoleMessage
Script source was shown.
Message added to source frame: test error message
Message removed from source frame: test error message

Running: testShowResource
